% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborateGSODdf.r
\name{elaborateGSODdf}
\alias{elaborateGSODdf}
\title{elaborateGSODdf}
\usage{
elaborateGSODdf(temp_daily)
}
\arguments{
\item{temp_daily}{Data.frame  Original raw data  file  from GSOD Global Daily SUmmary data.}
}
\value{
Return   data.frame Data are returned as R data.frame object.
}
\description{
Elaborate meteorological data from GSOD data.frames and return a data selection.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{,}
\keyword{GSOD.}
\keyword{data,daily}
\keyword{value}
\keyword{weather}

