% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborateGSODfiles.r
\name{elaborateGSODfiles}
\alias{elaborateGSODfiles}
\title{elaborateGSODfiles}
\usage{
elaborateGSODfiles(usaf, start_year = 1979, end_year = 2015,
  deletefile = F)
}
\arguments{
\item{usaf}{Character  USAF Code of file dowloaded from GSOD Global Daily SUmmary data.}

\item{start_year}{Numeric  Starting year of analisys.}

\item{end_year}{Numeric  Ending year of analisys.}
}
\value{
Return   data.frame Data are returned as R data.frame object.
}
\description{
Extract  meteorological data from GSOD gzip files and return a data selection.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{,}
\keyword{GSOD.}
\keyword{data,daily}
\keyword{value}
\keyword{weather}

