% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_weather.r
\name{get_daily_weather}
\alias{get_daily_weather}
\title{get_daily_weather}
\usage{
get_daily_weather(typestaz = "WU-STAZ", startdate, endate, idstaz = "LIRQ")
}
\arguments{
\item{typestaz}{Character Type of data "WU-STAZ" done by ICAO code of airport,  or "WU-PWS" by id of personal weather station.}

\item{startdate}{Character Initial date of period in "YYYY-MM-DD" format.}

\item{idstaz}{Character  Weather station ID or code for identification}

\item{enddate}{Character Final date of period in "YYYY-MM-DD" format.}
}
\value{
Return   data.frame Data are returned as R data.frame object.
}
\description{
Retrieve meteorological data from two different source : ICAO wunderground network data or PWS by using weatherData R packages.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{data,daily,wunderground,weatherData.}
\keyword{weather,}

