% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsod_data_arrange.r
\name{gsod_data_arrange}
\alias{gsod_data_arrange}
\title{gsod_data_arrange}
\usage{
gsod_data_arrange(gsod_staz)
}
\arguments{
\item{gsod_staz}{Character  Data.frame obtained with GSODToools R packages.}
}
\value{
Return   data.frame Data are returned as R data.frame object.
}
\description{
Arrange daily GSOD data in more reliable format.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{NOAA}
\keyword{data,daily,GSOD,}
\keyword{weather,}

