% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch2millimeter.r
\name{inch2millimeter}
\alias{inch2millimeter}
\title{inch2millimeter}
\usage{
inch2millimeter(val, ...)
}
\arguments{
\item{val}{Value in inches.}
}
\value{
Return   Numeric value in millimeter.
}
\description{
Return values in millimeter from inches. Is possible to set round. Function taken from GSODtools R package \url{https://zenodo.org/record/12217#.Vugi0V5VKlM}.
}
\author{
Florian Detsch
}
\references{
GSODtools R package \url{https://zenodo.org/record/12217#.Vugi0V5VKlM}.
}
\keyword{GSOD}
\keyword{data}

