% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2mps.r
\name{knots2mps}
\alias{knots2mps}
\title{knots2mps}
\usage{
knots2mps(val, decint = 2)
}
\arguments{
\item{val}{numeric: Value in inches.}

\item{decint}{integer: Decimal for rounding. Default is 2.}

\item{val}{numeric Velocity of wind in knots.}
}
\value{
Return   Numeric value in meterOnsecond.
}
\description{
Return  mps velocity from knots value.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{GSOD}
\keyword{data}

