% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2ms.r
\name{knots2ms}
\alias{knots2ms}
\title{knots2ms}
\usage{
knots2ms(val, ...)
}
\arguments{
\item{val}{Numeric Velocity of wind in meterstosecond.}
}
\value{
Return   Numeric
}
\description{
Return  ms velocity from knots value.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{GSOD}
\keyword{data}
