% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUCEAfile.r
\name{readUCEAfile}
\alias{readUCEAfile}
\title{readUCEAfile}
\usage{
readUCEAfile(file, param = "TEMPARIA2M_MAXG")
}
\arguments{
\item{file}{Character  Name of file dowloaded from http://cma.entecra.it/Banca_dati_agrometeo/index3.htm}

\item{param}{Character Type of name of parameter defined at http://cma.entecra.it/Banca_dati_agrometeo/miekal100_elepara_3.htm}
}
\value{
Return   data.frame Data are returned as R data.frame object.
}
\description{
Extract  meteorological data from UCEA \url{http://cma.entecra.it/Banca_dati_agrometeo/} gzip daily data file.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{data,daily,UCEA}
\keyword{weather,}

