% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveGSOD.r
\name{retrieveGSOD}
\alias{retrieveGSOD}
\title{retrieveGSOD}
\usage{
retrieveGSOD(usaf, WBAN = "99999", start_year = NA, end_year = NA,
  dsn = ".")
}
\arguments{
\item{usaf}{character  USAF code of GSOD Global Summary of Daily climatic data.}

\item{WBAN}{character  Another string for file naming.}

\item{start_year}{numeric Starting Year of data retrieval.}

\item{end_year}{numeric  Ending Year of data retrieval.}

\item{dsn}{character  Path of retrieved data}
}
\value{
Return   data.frame Data are returned as R data.frame object.
}
\description{
Download meteorological data from GSOD as  gzip files form.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{,}
\keyword{GSOD.}
\keyword{daily}
\keyword{data,}
\keyword{value}
\keyword{weather}

