% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_cfr_sp.r
\docType{data}
\name{station_cfr_sp}
\alias{station_cfr_sp}
\title{Dataset station_cfr_sp}
\format{SpatialPointsDataframe   Position of weather stations of Regione Toscana.}
\source{
\url{http://dati.toscana.it/dataset/stazioni-meteo-idrologiche/}
}
\usage{
station_cfr_sp
}
\description{
Dataset station_cfr_sp
}
\references{
Regione Toscana CRF Centro regionale Funzionale di Monitoraggio Meteo Meteorologico \email{http://www.sir.toscana.it/}
}
\keyword{data,opendata,regione}
\keyword{toscana}
\keyword{weather,}

