% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_gsod_ita.r
\docType{data}
\name{station_gsod_ita}
\alias{station_gsod_ita}
\title{Dataset station_gsod_ita}
\format{data.frame   Informations of weather stations of Italian NOAA GSOD.}
\source{
\url{http://www7.ncdc.noaa.gov/CDO/cdoselect.cmd?datasetabbv=GSOD}
}
\usage{
station_gsod_ita
}
\description{
Dataset station_gsod_ita
}
\references{
Global Summary of the Day (GSOD) please see Data Policy WM0 Resolution 40.
}
\keyword{datasets}

