% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_gsod_sp.r
\docType{data}
\name{station_gsod_sp}
\alias{station_gsod_sp}
\title{Dataset station_gsod_sp}
\format{SpatialPointsDataframe   Locations of weather stations of NOAA GSOD.}
\source{
\url{http://www7.ncdc.noaa.gov/CDO/cdoselect.cmd?datasetabbv=GSOD}
}
\usage{
station_gsod_sp
}
\description{
Dataset station_gsod_sp
}
\references{
Global Summary of the Day (GSOD) please see Data Policy WM0 Resolution 40.
}
\keyword{datasets}

