% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_sir_sp.r
\docType{data}
\name{station_sir_sp}
\alias{station_sir_sp}
\title{Dataset station_sir_sp}
\format{SpatialPointsDataframe   Position of weather stations of Regione Toscana.}
\source{
\url{http://www.sir.toscana.it/}
}
\usage{
station_sir_sp
}
\description{
Dataset station_sir_sp
}
\references{
Regione Toscana CRF Centro regionale Funzionale di Monitoraggio Meteo Meteorologico \email{http://www.sir.toscana.it/}
}
\keyword{data,opendata,regione}
\keyword{toscana}
\keyword{weather,}

