% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_wunderground_sp.r
\docType{data}
\name{station_wunderground_sp}
\alias{station_wunderground_sp}
\title{Dataset station_wunderground_sp}
\format{SpatialPointsDataframe   Locations of weather stations of Italian Personal Weather stations.}
\source{
\url{https://www.wunderground.com/weatherstation/overview.asp}
}
\usage{
station_wunderground_sp
}
\description{
Dataset station_wunderground_sp
}
\references{
Global weather network https://www.wunderground.com/
}
\keyword{datasets}

