% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendList.r
\name{appendlist}
\alias{appendlist}
\title{appendList}
\usage{
appendlist(x, val)
}
\value{
Return   A data.frame of xts object.
}
\description{
Utility function to append element to a list.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{data,opendata,regione}
\keyword{toscana}
\keyword{weather,}

