% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frame_TOS_cfr.r
\name{get_frame_TOS_cfr}
\alias{get_frame_TOS_cfr}
\title{get_frame_TOS_cfr}
\usage{
get_frame_TOS_cfr(query = "json_stations")
}
\arguments{
\item{query}{Query to retrieve framework.}
}
\value{
Return  data.fame Data from http://www.sir.toscana.it/archivio/
}
\description{
Retrieve framework data from CRF Centro regionale Funzionale di Monitoraggio Meteo Meteorologico of Regione Toscana.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\references{
Regione Toscana CRF Centro regionale Funzionale di Monitoraggio Meteo Meteorologico \email{http://www.sir.toscana.it/}
}
\keyword{data,opendata,regione}
\keyword{toscana}
\keyword{weather,}

