% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch2millimeter.r
\name{inch2millimeter}
\alias{inch2millimeter}
\title{inch2millimeter}
\usage{
inch2millimeter(val, decint = 2)
}
\arguments{
\item{val}{numeric: Value in inches.}

\item{decint}{integer: Decimal for rounding. Default is 2.}
}
\value{
Return   numeric Value in millimeter.
}
\description{
Return values in millimeter from inches. Is possible to set round. Function taken from GSODtools R package \url{https://zenodo.org/record/12217#.Vugi0V5VKlM}.
}
\author{
Alfonso Crisci.
}
\references{
GSODtools R package \url{https://zenodo.org/record/12217#.Vugi0V5VKlM}.
}
\keyword{GSOD}
\keyword{data}

