% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2mps.r
\name{knots2mps}
\alias{knots2mps}
\title{knots2mps}
\usage{
knots2mps(val, ...)
}
\arguments{
\item{val}{Numeric Velocity of wind in knots.}
}
\value{
Return   Numeric
}
\description{
Return  mps velocity from knots value.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{GSOD}
\keyword{data}

