% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knots2ms.r
\name{knots2ms}
\alias{knots2ms}
\title{knots2ms}
\usage{
knots2ms(val, decint = 2)
}
\arguments{
\item{val}{numeric: Value in knots.}

\item{decint}{integer: Decimal for rounding. Default is 2.}

\item{val}{numeric Velocity of wind in meteronsecond.}
}
\value{
Return   numeric
}
\description{
Return  metertoseconds velocity from knots value.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{GSOD}
\keyword{data}

