% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts2dfstaz.r
\name{xts2dfstaz}
\alias{xts2dfstaz}
\title{xts2dfstaz}
\usage{
xts2dfstaz(x, name, type = "pluvio")
}
\value{
Return   A data.frame of xts object.
}
\description{
Write and return data.frame object of xts object
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso crisci \email{a.crisci@ibimet.cnr.it}.
}
\keyword{data,xts}

